change radius variable type to integer

+++ include/Particle.h
@@ -5,11 +5,11 @@
 class Particle {
   Vector position_;
   Vector velocity_;
-  double radius_;
+  int radius_;
   double mass_;
 
  public:
-  Particle(const Vector& position, const Vector& velocity, double radius, double mass);
+  Particle(const Vector& position, const Vector& velocity, int radius, double mass);
 
   Vector GetImpulse();
